qr_code_generator/requirements.txt
Flask==3.0.3
rcode[pil]==7.4.2
